refactor: extract 3 chart components from main.tsx (Block B)
