+++ web/console/src/main.tsx
@@ -5,7 +5,6 @@ import { useUIStore } from './store/useUIStore';
 import { useTradingStore } from './store/useTradingStore';
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import ReactDOM from "react-dom/client";
-import { Time, CandlestickSeries, ColorType, CrosshairMode, LineStyle, createChart, createSeriesMarkers } from "lightweight-charts";
 import "./styles.css";
 
 import { API_BASE, fetchJSON } from './utils/api';
@@ -17,6 +16,9 @@ import { MetricCard } from './components/ui/MetricCard';
 import { FilterGroup } from './components/ui/FilterGroup';
 import { SimpleTable } from './components/ui/SimpleTable';
 import { SampleCard } from './components/ui/SampleCard';
+import { TradingChart } from './components/charts/TradingChart';
+import { SignalBarChart } from './components/charts/SignalBarChart';
+import { SignalMonitorChart } from './components/charts/SignalMonitorChart';
 
 import { AccountSummary, AccountRecord, StrategyVersion, StrategyRecord, AccountEquitySnapshot, Order, Fill, Position, PaperSession, LiveSession, ChartCandle, ChartAnnotation, MarkerLegendItem, BacktestRun, BacktestOptions, LiveAdapter, SignalSourceDefinition, SignalSourceCatalog, SignalSourceType, SignalBinding, SignalRuntimeAdapter, SignalRuntimeSession, ReplayReasonStats, ReplaySample, ExecutionTrade, SourceFilter, EventFilter, TimeWindow, MarkerDetail, ChartOverrideRange, SelectedSample, SelectableSample, RuntimeMarketSnapshot, RuntimeSourceSummary, RuntimeReadiness, SignalBarCandle, AlertItem, PlatformAlert, PlatformNotification, TelegramConfig, RuntimePolicy, LivePreflightSummary, LiveNextAction, LiveDispatchPreview, LiveSessionExecutionSummary, LiveSessionHealth, HighlightedLiveSession, LiveSessionFlowStep, SessionMarker, AuthSession } from './types/domain';
 import { sampleStatus, buildLinePath, summarizeRange, summarizeTimeRange, filterChartAnnotations, matchesEventFilter, resolveChartAnchor, buildTimeRange, buildSampleRange, buildSampleKey, annotationMatchesSample, findNearestAnnotation, toMarkerDetail, markerShape, markerPosition, markerColor, markerText, annotationTone, paperAccountsFromSummaries, strategyLabel, getNumber, getRecord, getList, deriveRuntimeMarketSnapshot, deriveRuntimeSourceSummary, deriveRuntimeReadiness, deriveSignalBarCandles, mapChartCandlesToSignalBarCandles, applyDefaultChartWindow, derivePrimarySignalBarState, buildRuntimeEventNotes, buildSourceStateNotes, buildSignalBarDecisionNotes, buildSignalBarStateNotes, deriveSignalActionSummary, deriveLivePreflightSummary, deriveLiveDispatchPreview, deriveLiveSessionExecutionSummary, derivePaperSessionExecutionSummary, deriveSessionMarkers, deriveLiveSessionHealth, deriveHighlightedLiveSession, deriveLiveSessionFlow, liveSessionHealthPriority, deriveLiveNextAction, liveSessionHealthTone, buildSignalActionNotes, buildTimelineNotes, summarizeOrderPreflight, derivePaperAlerts, deriveLiveAlerts, dedupeAlerts, buildAlertNotes, alertLevelTone, alertScopeTone, telegramDeliveryTone, runtimeReadinessTone, decisionStateTone, signalKindTone, signalActionTone, boolTone, boolLabel } from './utils/derivation';
@@ -4404,275 +4406,6 @@ function App() {
     </>
 
   );
-}
-
-function TradingChart(props: {
-  candles: ChartCandle[];
-  annotations: ChartAnnotation[];
-  focusTime?: string;
-  focusNonce: number;
-  selectedAnnotationIds: string[];
-  onSelectAnnotation: (annotation: ChartAnnotation) => void;
-  onHoverMarker: (detail: MarkerDetail | null) => void;
-}) {
-  const containerRef = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    if (!containerRef.current || props.candles.length === 0) {
-      return;
-    }
-
-    const chart = createChart(containerRef.current, {
-      autoSize: true,
-      height: 360,
-      layout: {
-        background: { type: ColorType.Solid, color: "rgba(255, 251, 242, 0.24)" },
-        textColor: "#4f585d",
-      },
-      grid: {
-        vertLines: { color: "rgba(216, 207, 186, 0.35)", style: LineStyle.Dotted },
-        horzLines: { color: "rgba(216, 207, 186, 0.35)", style: LineStyle.Dotted },
-      },
-      crosshair: {
-        mode: CrosshairMode.Normal,
-      },
-      rightPriceScale: {
-        borderColor: "rgba(216, 207, 186, 0.9)",
-      },
-      timeScale: {
-        borderColor: "rgba(216, 207, 186, 0.9)",
-        timeVisible: true,
-        secondsVisible: false,
-      },
-    });
-
-    const series = chart.addSeries(CandlestickSeries, {
-      upColor: "#0e6d60",
-      downColor: "#b04a37",
-      wickUpColor: "#0e6d60",
-      wickDownColor: "#b04a37",
-      borderVisible: false,
-      priceLineVisible: true,
-    });
-
-    series.setData(
-      props.candles.map((item) => ({
-        time: Math.floor(new Date(item.time).getTime() / 1000) as Time,
-        open: item.open,
-        high: item.high,
-        low: item.low,
-        close: item.close,
-      }))
-    );
-
-    const markers = createSeriesMarkers(
-      series,
-      props.annotations.map((item) => ({
-        time: Math.floor(new Date(item.time).getTime() / 1000) as Time,
-        position: markerPosition(item.type),
-        color: markerColor(item, props.selectedAnnotationIds.includes(item.id)),
-        shape: markerShape(item.type),
-        text: markerText(item, props.selectedAnnotationIds.includes(item.id)),
-      }))
-    );
-    markers.setMarkers(
-      props.annotations.map((item) => ({
-        time: Math.floor(new Date(item.time).getTime() / 1000) as Time,
-        position: markerPosition(item.type),
-        color: markerColor(item, props.selectedAnnotationIds.includes(item.id)),
-        shape: markerShape(item.type),
-        text: markerText(item, props.selectedAnnotationIds.includes(item.id)),
-      }))
-    );
-
-    chart.subscribeCrosshairMove((param) => {
-      if (param.time == null) {
-        props.onHoverMarker(null);
-        return;
-      }
-      const hoveredTime = Number(param.time);
-      if (!Number.isFinite(hoveredTime)) {
-        props.onHoverMarker(null);
-        return;
-      }
-
-      const nearest = findNearestAnnotation(props.annotations, hoveredTime);
-      props.onHoverMarker(nearest ? toMarkerDetail(nearest) : null);
-    });
-
-    chart.subscribeClick((param) => {
-      if (param.time == null) {
-        return;
-      }
-      const clickedTime = Number(param.time);
-      if (!Number.isFinite(clickedTime)) {
-        return;
-      }
-      const nearest = findNearestAnnotation(props.annotations, clickedTime);
-      if (nearest) {
-        props.onSelectAnnotation(nearest);
-      }
-    });
-
-    if (props.focusTime && props.focusNonce > 0) {
-      const focusSeconds = Math.floor(new Date(props.focusTime).getTime() / 1000);
-      const firstSeconds = Math.floor(new Date(props.candles[0].time).getTime() / 1000);
-      const lastSeconds = Math.floor(new Date(props.candles[props.candles.length - 1].time).getTime() / 1000);
-      const span = Math.max(lastSeconds - firstSeconds, 60 * 60);
-      const padding = Math.max(Math.floor(span / 6), 30 * 60);
-      chart.timeScale().setVisibleRange({
-        from: (focusSeconds - padding) as Time,
-        to: (focusSeconds + padding) as Time,
-      });
-    } else {
-      chart.timeScale().fitContent();
-    }
-    return () => {
-      props.onHoverMarker(null);
-      chart.remove();
-    };
-  }, [
-    props.annotations,
-    props.candles,
-    props.focusNonce,
-    props.focusTime,
-    props.onHoverMarker,
-    props.onSelectAnnotation,
-    props.selectedAnnotationIds,
-  ]);
-
-  return <div ref={containerRef} className="tv-chart" />;
-}
-
-function SignalBarChart(props: { candles: SignalBarCandle[] }) {
-  const containerRef = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    if (!containerRef.current || props.candles.length === 0) {
-      return;
-    }
-
-    const chart = createChart(containerRef.current, {
-      autoSize: true,
-      height: 260,
-      layout: {
-        background: { type: ColorType.Solid, color: "rgba(255, 251, 242, 0.16)" },
-        textColor: "#4f585d",
-      },
-      grid: {
-        vertLines: { color: "rgba(216, 207, 186, 0.25)", style: LineStyle.Dotted },
-        horzLines: { color: "rgba(216, 207, 186, 0.25)", style: LineStyle.Dotted },
-      },
-      crosshair: { mode: CrosshairMode.Normal },
-      rightPriceScale: { borderColor: "rgba(216, 207, 186, 0.9)" },
-      timeScale: {
-        borderColor: "rgba(216, 207, 186, 0.9)",
-        timeVisible: true,
-        secondsVisible: false,
-      },
-    });
-
-    const series = chart.addSeries(CandlestickSeries, {
-      upColor: "#0e6d60",
-      downColor: "#b04a37",
-      wickUpColor: "#0e6d60",
-      wickDownColor: "#b04a37",
-      borderVisible: false,
-      priceLineVisible: true,
-    });
-
-    series.setData(
-      props.candles.map((item) => ({
-        time: Math.floor(new Date(item.time).getTime() / 1000) as Time,
-        open: item.open,
-        high: item.high,
-        low: item.low,
-        close: item.close,
-      }))
-    );
-
-    applyDefaultChartWindow(chart, props.candles.length, 90);
-    return () => chart.remove();
-  }, [props.candles]);
-
-  return <div ref={containerRef} className="tv-chart" />;
-}
-
-function SignalMonitorChart(props: { candles: SignalBarCandle[]; markers: SessionMarker[] }) {
-  const containerRef = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    if (!containerRef.current || props.candles.length === 0) {
-      return;
-    }
-
-    const chart = createChart(containerRef.current, {
-      autoSize: true,
-      height: 360,
-      layout: {
-        background: { type: ColorType.Solid, color: "rgba(255, 251, 242, 0.20)" },
-        textColor: "#4f585d",
-      },
-      grid: {
-        vertLines: { color: "rgba(216, 207, 186, 0.25)", style: LineStyle.Dotted },
-        horzLines: { color: "rgba(216, 207, 186, 0.25)", style: LineStyle.Dotted },
-      },
-      crosshair: { mode: CrosshairMode.Normal },
-      rightPriceScale: { borderColor: "rgba(216, 207, 186, 0.9)" },
-      timeScale: {
-        borderColor: "rgba(216, 207, 186, 0.9)",
-        timeVisible: true,
-        secondsVisible: false,
-        barSpacing: 10,
-        minBarSpacing: 4,
-        rightOffset: 6,
-      },
-    });
-
-    const series = chart.addSeries(CandlestickSeries, {
-      upColor: "#0e6d60",
-      downColor: "#b04a37",
-      wickUpColor: "#0e6d60",
-      wickDownColor: "#b04a37",
-      borderVisible: false,
-      priceLineVisible: true,
-    });
-
-    series.setData(
-      props.candles.map((item) => ({
-        time: Math.floor(new Date(item.time).getTime() / 1000) as Time,
-        open: item.open,
-        high: item.high,
-        low: item.low,
-        close: item.close,
-      }))
-    );
-
-    const markers = createSeriesMarkers(
-      series,
-      props.markers.map((item) => ({
-        time: Math.floor(new Date(item.time).getTime() / 1000) as Time,
-        position: item.position,
-        color: item.color,
-        shape: item.shape,
-        text: item.text,
-      }))
-    );
-    markers.setMarkers(
-      props.markers.map((item) => ({
-        time: Math.floor(new Date(item.time).getTime() / 1000) as Time,
-        position: item.position,
-        color: item.color,
-        shape: item.shape,
-        text: item.text,
-      }))
-    );
-
-    applyDefaultChartWindow(chart, props.candles.length, 90);
-    return () => chart.remove();
-  }, [props.candles, props.markers]);
-
-  return <div ref={containerRef} className="tv-chart" />;
 }
